pyproject.toml
[tool.poetry]
name = "shawarmaasaservice_bot"
version = "0.1.0"
description = ""
authors = ["Vladimir Turov <firemoon@icloud.com>"]
license = "MIT"
packages = [
    { include = "alembic_migrations" },
    { include = "shaas_bot" },
    { include = "shaas_web" },
    { include = "shaas_common" },
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.78.0"
sqlalchemy = "1.4.39"
asyncpg = "0.25.0"
uvicorn = "0.18.2"
python-telegram-bot = "20.0.0a0"
toml = "0.10.2"
jinja2 = "3.1.2"
python-multipart = "0.0.5"
lazy = "1.4"

[tool.poetry.dev-dependencies]
alembic="1.8.0"
psycopg2-binary="2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
